refactor: 상담기록 외래키 추가 | 추가,상세보기, 삭제, 수정에서 현재 멤버 소유인지 확인

+++ src/main/java/solverz/business_card/domain/card/service/CardService.java
@@ -88,6 +88,9 @@ public class CardService {
     public DeleteCardResponse deleteCard(DeleteCardRequest request) {
         Card card = cardRepository.findById(request.cardId())
                 .orElseThrow(() -> new BusinessException(ErrorCode.DELETION_FAILED_CARD));
+        if (!card.getMember().getMemberToken().equals(request.memberToken())) {
+            throw new BusinessException(ErrorCode.NOT_CARD_OWNER);
+        }
         cardRepository.deleteById(request.cardId());
         return DeleteCardResponse.from(card);
     }
